+++ src/ResourceManager/TrafficManager/ChangeLog.md
@@ -21,6 +21,7 @@
 * Support for the Geographic traffic routing method
     - New value 'Geographic' for the TrafficRoutingMethod parameter of New-AzureRmTrafficManagerProfile
     - New parameter 'GeoMapping' for the New-AzureRmTrafficManagerEndpoint and Add-AzureRmTrafficManagerEndpointConfig
+    - Fix piping for Get-AzureRmTrafficManagerProfile when it returns a collection of profiles
 
 ## Version 2.7.0